refactor: rename StepIndicatorProps.ts to defaults.ts
The file contains stepIndicatorDefaultProps (runtime default values), not
type definitions. The old name was misleading. Renamed to defaults.ts to
accurately describe the file's content and align with conventional naming.
