+++ packages/dnb-eufemia/src/components/step-indicator/defaults.ts
